+++ mne_bids/config.py
@@ -65,7 +65,8 @@ ALLOWED_MODALITY_EXTENSIONS = {'meg': allowed_extensions_meg,
                                'eeg': allowed_extensions_eeg,
                                'ieeg': allowed_extensions_ieeg}
 
-# allowed kinds (i.e. last "_" delimiter in the BIDS filenames)
+# allowed suffixes (i.e. last "_" delimiter in the BIDS filenames before
+# the extension)
 ALLOWED_FILENAME_SUFFIX = [
     'meg', 'markers', 'eeg', 'ieeg', 'T1w',  # modality
     'participants', 'scans',