+++ mne_bids/tests/test_utils.py
@@ -43,19 +43,19 @@ def test_get_ch_type_mapping():
         _get_ch_type_mapping(fro='bogus', to='mne')
 
 
-def test_handle_kind():
+def test_handle_modality():
     """Test the automatic extraction of str_suffix from the data."""
     # Create a dummy raw
     n_channels = 1
     sampling_rate = 100
     data = random((n_channels, sampling_rate))
     channel_types = ['grad', 'eeg', 'ecog']
-    expected_kinds = ['meg', 'eeg', 'ieeg']
+    expected_modalities = ['meg', 'eeg', 'ieeg']
     # do it once for each type ... and once for "no type"
-    for chtype, kind in zip(channel_types, expected_kinds):
+    for chtype, modality in zip(channel_types, expected_modalities):
         info = mne.create_info(n_channels, sampling_rate, ch_types=[chtype])
         raw = mne.io.RawArray(data, info)
-        assert _handle_modality(raw) == kind
+        assert _handle_modality(raw) == modality
 
     # if the situation is ambiguous (EEG and iEEG channels both), raise error
     with pytest.raises(ValueError, match='Both EEG and iEEG channels found'):