+++ mne_bids/utils.py
@@ -88,22 +88,22 @@ def _get_ch_type_mapping(fro='mne', to='bids'):
 
 
 def _handle_modality(raw):
-    """Get kind."""
+    """Get modality."""
     if 'eeg' in raw and ('ecog' in raw or 'seeg' in raw):
         raise ValueError('Both EEG and iEEG channels found in data.'
                          'There is currently no specification on how'
                          'to handle this data. Please proceed manually.')
     elif 'meg' in raw:
-        kind = 'meg'
+        modality = 'meg'
     elif 'ecog' in raw or 'seeg' in raw:
-        kind = 'ieeg'
+        modality = 'ieeg'
     elif 'eeg' in raw:
-        kind = 'eeg'
+        modality = 'eeg'
     else:
         raise ValueError('Neither MEG/EEG/iEEG channels found in data.'
                          'Please use raw.set_channel_types to set the '
                          'channel types in the data.')
-    return kind
+    return modality
 
 
 def _age_on_date(bday, exp_date):